feat: adicionar exception ao método deletar endereco no endereco_service

+++ src/main/java/cepein/mapeamento/endereco/service/EnderecoService.java
@@ -55,7 +55,11 @@ public class EnderecoService {
     }
 
     @Transactional
-    public ResponseEntity<Void> deletarEndereco(Long idEndereco){
+    public Void deletarEndereco(Long idEndereco){
+        Endereco endereco = this.buscarEndereco(idEndereco);
+        if(endereco.getPessoaPorId() != null && endereco.getPessoaPorUuid() != null){
+            throw new PessoaExistException("Uma pessoa possui este endereço, não é possivel deleta-lo");
+        }
         this.enderecoRepository.deleteById(idEndereco);
         return null;
     }
